requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
python-dotenv
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
